usa-car-finder/requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic==2.7.1
fastapi==0.111.0
uvicorn==0.29.0
anthropic==0.28.0
jinja2==3.1.4
weasyprint==62.3
reportlab==4.4.10
python-dotenv==1.0.1
httpx==0.27.0
aiohttp
